Attempts to pass QWidget as parameter
WIP
